% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModel.R, R/misc.R,
%   R/predictionFunction.R, R/predictors.R, R/probFunction.R,
%   R/selectByFilter.R, R/sortImp.R, R/workflows.R
\name{caret-internal}
\alias{caret-internal}
\alias{createModel}
\alias{sortImp}
\alias{caretTheme}
\alias{progress}
\alias{hasTerms}
\alias{predictionFunction}
\alias{probFunction}
\alias{expandParameters}
\alias{flatTable}
\alias{MeanSD}
\alias{sbfIter}
\alias{gamFormula}
\alias{bagEarthStats}
\alias{cforestStats}
\alias{ipredStats}
\alias{rfStats}
\alias{well_numbered}
\alias{outcome_conversion}
\title{Internal Functions}
\usage{
createModel(
  x,
  y,
  wts,
  method,
  tuneValue,
  obsLevels,
  pp = NULL,
  last = FALSE,
  sampling = NULL,
  classProbs,
  ...
)

well_numbered(prefix, items)

gamFormula(data, smoother = "s", cut = 8, y = "y")

flatTable(pred, obs)

ipredStats(x)

rfStats(x)

cforestStats(x)

bagEarthStats(x)

outcome_conversion(x, lv)

predictionFunction(method, modelFit, newdata, preProc = NULL, param = NULL)

hasTerms(x)

probFunction(method, modelFit, newdata = NULL, preProc = NULL, param = NULL)

sbfIter(x, y, testX, testY, testPerf = NULL, sbfControl = sbfControl(), ...)

sortImp(object, top)

progress(x, names, iter, start = TRUE)

MeanSD(x, exclude = NULL)

expandParameters(fixed, seq)
}
\description{
Internal functions
}
\author{
Max Kuhn, but \code{caretTheme} uses an expanded grid of the "Blues"
palette designed by Cynthia Brewer and Mark Harrower
}
\keyword{internal}
