% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metric_transformed_data_viz_3d.R
\name{Cols}
\alias{Cols}
\title{Assigning Colors to A Vector}
\usage{
Cols(vec)
}
\arguments{
\item{vec}{The vector where each distinct value will be assigned a color.}
}
\value{
The colors for each element in the given vector
}
\description{
This function assigns a color to each distinct value in the given vector.
}

