% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret-package.R
\docType{data}
\name{GermanCredit}
\alias{GermanCredit}
\title{German Credit Data}
\source{
UCI Machine Learning Repository
}
\description{
Data from Dr. Hans Hofmann of the University of Hamburg.
}
\details{
These data have two classes for the credit worthiness: good or bad. There
are predictors related to attributes, such as: checking account status,
duration, credit history, purpose of the loan, amount of the loan, savings
accounts or bonds, employment duration, Installment rate in percentage of
disposable income, personal information, other debtors/guarantors, residence
duration, property, age, other installment plans, housing, number of
existing credits, job information, Number of people being liable to provide
maintenance for, telephone, and foreign worker status.

Many of these predictors are discrete and have been expanded into several
0/1 indicator variables
}
\keyword{datasets}
