% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getSamplingInfo}
\alias{getSamplingInfo}
\title{Get sampling info from a train model}
\usage{
getSamplingInfo(method = NULL, regex = TRUE, ...)
}
\arguments{
\item{method}{Modeling method.}

\item{regex}{Whether to use regex matching.}

\item{...}{additional arguments to passed to grepl.}
}
\value{
A list
}
\description{
Placeholder.
}
\details{
Placeholder.
}
