% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfda.R
\name{\%^\%}
\alias{\%^\%}
\title{Negative One Half Matrix Power Operator}
\usage{
x \%^\% n
}
\arguments{
\item{x}{the matrix we want to operate on}

\item{n}{the exponent}
}
\value{
the matrix after negative one half power
}
\description{
This function defines operation for negative one half matrix
power operator
}

