% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfda.R
\name{lfda.default}
\alias{lfda.default}
\title{This function trains a lfda metric, returns a transformed original matrix and transforming matrix used to
transform other data set, usually for testing set}
\usage{
\method{lfda}{default}(x, y, r = 3, metric = c("orthonormalized", "plain",
  "weighted"), knn = 5, ...)
}
\description{
This function trains a lfda metric, returns a transformed original matrix and transforming matrix used to
transform other data set, usually for testing set
}

