% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelLookup.R
\name{modelLookup}
\alias{modelLookup}
\alias{getModelInfo}
\alias{checkInstall}
\title{Tools for Models Available in \code{train}}
\usage{
modelLookup(model = NULL)

checkInstall(pkg)

getModelInfo(model = NULL, regex = TRUE, ...)
}
\arguments{
\item{model}{a character string associated with the \code{method} argument of
\code{\link[=train]{train()}}. If no value is passed, all models are returned. For
\code{getModelInfo}, regular expressions can be used.}

\item{pkg}{a character string of package names.}

\item{regex}{a logical: should a regular expressions be used? If \code{FALSE}, a
simple match is conducted against the whole name of the model.}

\item{...}{options to pass to \code{\link[base:grepl]{base::grepl()}}}
}
\value{
\code{modelLookup} produces a data frame with columns:
\itemize{
\item \code{model}: a character string for the model code
\item \code{parameter}: the tuning parameter name
\item \code{label}: a tuning parameter label (used in plots)
\item \code{forReg}: a logical; can the model be used for regression?
\item \code{forClass}: a logical; can the model be used for classification?
\item \code{probModel}: a logical; does the model produce class probabilities?
}

\code{getModelInfo} returns a list containing one or more lists of the standard
model information. \code{checkInstall} returns not value.
}
\description{
These function show information about models and packages that are
accessible via \code{\link[=train]{train()}}
}
\details{
\code{modelLookup} is good for getting information related to the tuning
parameters for a model. \code{getModelInfo} will return all the functions and
metadata associated with a model. Both of these functions will only search
within the models bundled in this package.

\code{checkInstall} will check to see if packages are installed. If they are not
and the session is interactive, an option is given to install the packages
using \code{\link[utils:install.packages]{utils::install.packages()}} using that functions default arguments
(the missing packages are listed if you would like to install them with
other options). If the session is not interactive, an error is thrown.
}
\note{
The column \code{seq} is no longer included in the output of \code{modelLookup}.
}
\examples{
\dontshow{if (!caret:::is_cran_check()) withAutoprint(\{ # examplesIf}

\dontrun{
if (caret:::has_packages("pls")) {
  modelLookup()
  modelLookup("gbm")

  getModelInfo("pls")
  getModelInfo("^pls")
  getModelInfo("pls", regex = FALSE)

  checkInstall(getModelInfo("pls")$library)
}
}
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=train]{train()}}, \code{\link[utils:install.packages]{utils::install.packages()}}, \code{\link[base:grepl]{base::grepl()}}
}
\author{
Max Kuhn
}
\keyword{utilities}
