% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfda.R
\name{plot.lfda}
\alias{plot.lfda}
\title{3D Visualization on Metric Transformed Data}
\usage{
plot.lfda(x, labels, cleanText=FALSE)
}
\arguments{
\item{x}{The lfda model object.}

\item{labels}{A list of class labels used for lfda training.}

\item{cleanText}{A boolean value to specify whether to make the labels in the plot cleaner (default: FALSE)}

\item{...}{Additional arguments}
}
\description{
This function performs 3D visualization for a trained lfda model. 
}
\seealso{
See \code{\link{lfda}} for the metric learning model. 
}

