\name{pottery}
\docType{data}
\alias{pottery}
\alias{potteryClass}
\title{Pottery from Pre-Classical Sites in Italy}
\description{
Measurements of 58 pottery samples.
}

\usage{data(pottery)}

\value{
  \item{pottery}{11 elemental composition measurements }
  \item{potteryClass}{factor of pottery type: black carbon containing bulks (A) and clayey (B)}
  }


\source{R. G. Brereton  (2003). \emph{Chemometrics: Data Analysis for the Laboratory and Chemical Plant}, pg. 261.
}

\keyword{datasets}  
