% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfda.R
\name{predict.lfda}
\alias{predict.lfda}
\title{LFDA Transformation}
\usage{
\method{predict}{lfda}(object, newdata = NULL, type = "raw", ...)
}
\arguments{
\item{object}{The result from lfda function, which contains a transformed data and a transforming
matrix that can be used for transforming testing set}

\item{newdata}{The data to be transformed}

\item{type}{The output type, in this case it defaults to "raw" since the output is a matrix}

\item{...}{Additional arguments}
}
\value{
the transformed matrix
}
\description{
This function transforms a data set, usually a testing set, using the trained LFDA metric
}
\author{
Yuan Tang
}

