% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfda.R
\name{repmat}
\alias{repmat}
\title{Matlab-Syntaxed Repmat}
\usage{
repmat(A, N, M)
}
\arguments{
\item{A}{original matrix to be used as copies}

\item{N}{the number of rows of tiling copies of A}

\item{M}{the number of columns of tiling copies of A}
}
\value{
matrix consisting of an N-by-M tiling copies of A
}
\description{
This function mimics the behavior and syntax of repmat() in Matlab
it generates a large matrix consisting of an N-by-M tiling copies of A
}

