\name{scat}
\docType{data}
\alias{scat}
\alias{scat_orig}
\title{Morphometric Data on Scat}
\description{
Reid (2105) collected data on animal feses in coastal California. The data consist of DNA verified species designations as well as fields related to the time and place of the collection and the scat itself. The data frame \code{scat_orig} contains  while \code{scat} contains data on the three main species. 
}

\usage{data(scat)}

\value{
  \item{scat_orig}{the entire data set in the Supplemental Materials}
  \item{scat}{data on the three main species}
  }

\source{Reid, R. E. B. (2015). A morphometric modeling approach to distinguishing among bobcat, coyote and gray fox scats. \emph{Wildlife Biology}, 21(5), 254-262
}

\keyword{datasets}  

