% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModel.R
\name{caret-internal}
\alias{MeanSD}
\alias{bagEarthStats}
\alias{caret-internal}
\alias{caretTheme}
\alias{cforestStats}
\alias{createModel}
\alias{expandParameters}
\alias{flatTable}
\alias{gamFormula}
\alias{hasTerms}
\alias{ipredStats}
\alias{predictionFunction}
\alias{probFunction}
\alias{progress}
\alias{resampleWrapper}
\alias{rfStats}
\alias{sbfIter}
\alias{sortImp}
\alias{well_numbered}
\title{Internal Functions}
\usage{
createModel(x, y, wts, method, tuneValue, obsLevels,
	        pp = NULL, last = FALSE, sampling = NULL, classProbs, ...)
predictionFunction(method, modelFit, newdata,
	               preProc = NULL, param = NULL)
probFunction(method, modelFit, newdata,
	         preProc = NULL, param = NULL)
flatTable(pred, obs)
expandParameters(fixed, seq)
MeanSD(x, exclude = NULL)
sortImp(object, top)
resampleWrapper(x, ind)
caretTheme()
ipredStats(x)
rfStats(x)
bagEarthStats(x)
cforestStats(x)
progress(x, names, iter, start = TRUE)
hasTerms(x)
sbfIter(x, y, testX, testY, sbfControl = sbfControl(), ...)
gamFormula(data, smoother = "s", cut = 8, y = "y")
well_numbered(prefix, items)
}
\description{
Internal functions
}
\author{
Max Kuhn, but \code{caretTheme} uses an expanded grid of the "Blues" palette designed by Cynthia Brewer and Mark Harrower
}

