% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModel.R
\name{caret-internal}
\alias{MeanSD}
\alias{bagEarthStats}
\alias{caret-internal}
\alias{caretTheme}
\alias{cforestStats}
\alias{createModel}
\alias{expandParameters}
\alias{flatTable}
\alias{gamFormula}
\alias{hasTerms}
\alias{ipredStats}
\alias{predictionFunction}
\alias{probFunction}
\alias{progress}
\alias{resampleWrapper}
\alias{rfStats}
\alias{sbfIter}
\alias{sortImp}
\alias{well_numbered}
\title{Internal Functions}
\usage{
createModel(x, y, wts, method, tuneValue, obsLevels, pp = NULL,
  last = FALSE, sampling = NULL, classProbs, ...)
}
\description{
Internal functions
}
\author{
Max Kuhn, but \code{caretTheme} uses an expanded grid of the "Blues" palette designed by Cynthia Brewer and Mark Harrower
}

