% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModel.R, R/misc.R, R/predictionFunction.R, R/predictors.R, R/probFunction.R, R/resampleWrapper.R, R/selectByFilter.R, R/sortImp.R, R/workflows.R
\name{caret-internal}
\alias{MeanSD}
\alias{bagEarthStats}
\alias{caret-internal}
\alias{caretTheme}
\alias{cforestStats}
\alias{createModel}
\alias{expandParameters}
\alias{flatTable}
\alias{gamFormula}
\alias{hasTerms}
\alias{ipredStats}
\alias{predictionFunction}
\alias{probFunction}
\alias{progress}
\alias{resampleWrapper}
\alias{rfStats}
\alias{sbfIter}
\alias{sortImp}
\alias{well_numbered}
\title{Internal Functions}
\usage{
createModel(x, y, wts, method, tuneValue, obsLevels, pp = NULL,
  last = FALSE, sampling = NULL, classProbs, ...)

well_numbered(prefix, items)

gamFormula(data, smoother = "s", cut = 8, y = "y")

flatTable(pred, obs)

ipredStats(x)

rfStats(x)

cforestStats(x)

bagEarthStats(x)

predictionFunction(method, modelFit, newdata, preProc = NULL, param = NULL)

hasTerms(x)

probFunction(method, modelFit, newdata = NULL, preProc = NULL,
  param = NULL)

resampleWrapper(x, ind)

sbfIter(x, y, testX, testY, sbfControl = sbfControl(), ...)

sortImp(object, top)

progress(x, names, iter, start = TRUE)

MeanSD(x, exclude = NULL)

expandParameters(fixed, seq)
}
\description{
Internal functions
}
\author{
Max Kuhn, but \code{caretTheme} uses an expanded grid of the "Blues" palette designed by Cynthia Brewer and Mark Harrower
}
\keyword{internal}

