% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret-package.R
\docType{data}
\name{dhfr}
\alias{dhfr}
\title{Dihydrofolate Reductase Inhibitors Data}
\source{
Sutherland, J.J. and Weaver, D.F. (2004). Three-dimensional
quantitative structure-activity and structure-selectivity relationships of
dihydrofolate reductase inhibitors, \emph{Journal of Computer-Aided
Molecular Design}, Vol. 18, pg. 309-331.
}
\value{
\item{dhfr}{data frame of chemical descriptors and the activity
values}
}
\description{
Sutherland and Weaver (2004) discuss QSAR models for dihydrofolate reductase
(DHFR) inhibition. This data set contains values for 325 compounds. For each
compound, 228 molecular descriptors have been calculated. Additionally, each
samples is designated as "active" or "inactive".
}
\details{
The data frame \code{dhfr} contains a column called \code{Y} with the
outcome classification. The remainder of the columns are molecular
descriptor values.
}
\keyword{datasets}
