% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfda.R
\name{lfda}
\alias{lfda}
\title{Local Fisher Discriminant Analysis}
\usage{
lfda(x, ...)
}
\arguments{
\item{x}{The data frame to be used to train the metric}

\item{...}{Additional arguments}
}
\description{
Local Fisher Discriminant Analysis
}

