% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret-package.R
\docType{data}
\name{pottery}
\alias{pottery}
\alias{potteryClass}
\title{Pottery from Pre-Classical Sites in Italy}
\source{
R. G. Brereton (2003). \emph{Chemometrics: Data Analysis for the
Laboratory and Chemical Plant}, pg. 261.
}
\value{
\item{pottery}{11 elemental composition measurements }
\item{potteryClass}{factor of pottery type: black carbon containing bulks
(A) and clayey (B)}
}
\description{
Measurements of 58 pottery samples.
}
\keyword{datasets}
