% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampleSummary.R
\name{resampleSummary}
\alias{resampleSummary}
\title{Summary of resampled performance estimates}
\usage{
resampleSummary(obs, resampled, index = NULL, keepData = TRUE)
}
\arguments{
\item{obs}{A vector (numeric or factor) of the outcome data}

\item{resampled}{For bootstrapping, this is either a matrix (for numeric
outcomes) or a data frame (for factors). For cross-validation, a vector is
produced.}

\item{index}{The list to index of samples in each cross-validation fold
(only used for cross-validation).}

\item{keepData}{A logical for returning the observed and predicted data.}
}
\value{
A list with: \item{metrics }{A vector of values describing the
bootstrap distribution.} \item{data }{A data frame or \code{NULL}. Columns
include \code{obs}, \code{pred} and \code{group} (for tracking
cross-validation folds or bootstrap samples)}
}
\description{
This function uses the out-of-bag predictions to calculate overall
performance metrics and returns the observed and predicted data.
}
\details{
The mean and standard deviation of the values produced by
\code{\link{postResample}} are calculated.
}
\examples{

resampleSummary(rnorm(10), matrix(rnorm(50), ncol = 5))

}
\seealso{
\code{\link{postResample}}
}
\author{
Max Kuhn
}
\keyword{utilities}
