% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret-package.R
\docType{data}
\name{scat}
\alias{scat}
\alias{scat_orig}
\title{Morphometric Data on Scat}
\source{
Reid, R. E. B. (2015). A morphometric modeling approach to
distinguishing among bobcat, coyote and gray fox scats. \emph{Wildlife
Biology}, 21(5), 254-262
}
\value{
\item{scat_orig}{the entire data set in the Supplemental Materials}
\item{scat}{data on the three main species}
}
\description{
Reid (2015) collected data on animal feses in coastal California. The data
consist of DNA verified species designations as well as fields related to
the time and place of the collection and the scat itself. The data frame
\code{scat_orig} contains while \code{scat} contains data on the three main
species.
}
\keyword{datasets}
